requirements.txt
openai # Version: 1.76.0
python-dotenv
streamlit